requirements.txt
Django==1.6.5
Markdown==2.5.1
Twisted==14.0.0
argparse==1.2.1
cffi==0.8.6
characteristic==14.1.0
cryptography==0.5.4
datadiff==1.1.5
defusedxml==0.4.1
distribute==0.6.24
django-allauth==0.17.0
django-countries==2.1.2
django-datetime-widget==0.9.3
django-extensions==1.3.8
django-filter==0.8
django-ipware==0.1.0
django-jquery==1.9.1
django-jsonfield==0.9.13
django-jsonview==0.4.2
django-leaflet==0.13.7
django-passwords==0.2.0
django-periodically==0.3.0
django-rest-auth==0.3.2
django-selenium==0.9.6
django-session-security==2.2.0
django-user-sessions==1.1.0
djangorestframework==2.4.3
djorm-pgarray==1.1
eventlog==0.8.0
geoip2==2.1.0
ipaddr==2.1.11
maxminddb==1.1.1
numpy==1.9.1
oauthlib==0.6.3
psycopg2==2.5.3
pusher==0.8
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pyephem==3.7.5.3
pyserial==2.7
python-dateutil==2.2
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
rpc4django==0.3.0
service-identity==14.0.0
simplekml==1.2.3
six==1.8.0
wsgiref==0.1.2
zope.interface==4.1.1
